Duplicate checking
Editing the last large update text (for the todo mostly)  
ADDED:
Duplicate checker utility. Uses the system config and gives a list of duplicated cells.
Comments! lots of comments, everything is now full commented except for the function code in the gui file..

All current dependencies should someone want to run this:

    -Python 2.7.x
    -appJar (pip install appjar)
    -shutils, which may be builtin
    -tkinter
    -os, which is definitely builtin

To do:
   Important:
C O M M E N T
    * need to comment things in RENAMER-GUI detail, remove redundant code in the duplicate checker,etc
     * Check for file not found! This can cause the program to crash if you pick the wrong directory instead of tossing a warning.
     * Check for generic errors in anythin not gui related... such as validating configuration when you apply/ok, since otherwise you can currently enter any text/characters into the textboxes. (check if offset is number, check if columns are only letters with no spaces/special characters/numbers. (will cause renstuff to throw an error, ugly)

The above config validation will only make sure the entries are possibly valid, but without data validation it cant check if the configuration is actually correct.
    
    Less Important:
     simple data validation.. check the two picture number columns for repeated numbers...
If there is repeated numbers, report which two cells are repeated, and whether to cancel or continue.
before that , the program would check that there are only numbers in the cells..

    Very unimportant:
  
   A progress bar based on the index of total rows in the spreadsheet would be nice, if not particularly critical.